Moved 'AuthButton' code from Sidebar.tsx and over to AuthButton.tsx (overwrote original file for compatibility).
Now, there shouldn't be any duplicate 'AuthButton' code.
