+++ src/components/Sidebar/AuthButton.tsx
@@ -1,29 +1,31 @@
 'use client';
 
-import { useAuth } from '@/hooks/useAuth';
+import SidebarIcon from './SidebarIcon';
 
-export function AuthButton() {
-  const { user, loading, signOut, signIn, isAuthenticated } = useAuth();
-
+const AuthButton = ({
+  action,
+  label,
+}: {
+  action: () => void;
+  label: 'Sign In' | 'Sign Out';
+}) => {
   return (
-    <div className="flex items-center gap-3">
-      {isAuthenticated ? (
-        <span>Welcome, {user.name || user.email}!</span>
-      ) : null}
-
-      <button
-        className="bg-blue-300 hover:bg-blue-400 px-3 py-1 rounded-md cursor-pointer"
-        disabled={loading}
-        onClick={loading ? undefined : isAuthenticated ? signOut : signIn}
-      >
-        {loading ? (
-          <>Loading...</>
-        ) : isAuthenticated ? (
-          <>Sign Out</>
-        ) : (
-          <>Login with GitHub</>
-        )}
+    <li>
+      <button className="w-full justify-between" onClick={action}>
+        <SidebarIcon label={label} />
+        {label}
       </button>
-    </div>
+    </li>
   );
-}
+};
+
+export default AuthButton;
+
+
+
+
+
+
+
+
+
